Binary conversion data type changed

+++ Function_Sin_Flags.c
@@ -71,7 +71,7 @@ int Print_Number_Binario(va_list Lista)
 {
 	int Bits;
 
-	Bits = Print_Binario(va_arg(Lista, unsigned int), 0);
+	Bits = Print_Binario((long int)va_arg(Lista, unsigned int), 0);
 
 	return (Bits);
 }
